+++ forui/lib/src/widgets/popover/popover.dart
@@ -572,7 +572,7 @@ class FPopoverStyle with Diagnosticable, _$FPopoverStyleFunctions {
   FPopoverStyle.inherit({required FColors colors, required FStyle style})
     : this(
         decoration: BoxDecoration(
-          color: colors.background,
+          color: colors.popover,
           borderRadius: style.borderRadius,
           border: .all(width: style.borderWidth, color: colors.border),
           boxShadow: style.shadow,